Initial footer and left-nav implementation.
