+++ webapp/src/app/layout/app-container/app-container.component.html
@@ -1,16 +1,6 @@
 <div class="app-container">
-    <div class="navigation">
-        <!-- TODO: Add navigation -->
-        (nav)
-        <p><sbdl-button class ="primary" [routerLink]="['']">Home</sbdl-button></p>
-        <p><sbdl-button class ="primary" [routerLink]="['/resource/1']">IR</sbdl-button></p>
-        <p><sbdl-button class ="primary" [routerLink]="['/resource/3']">PL</sbdl-button></p>
-        <p><sbdl-button class ="primary" *ngIf="!isAuthenticated" [routerLink]="['/auth/login']">Lgin</sbdl-button></p>
-        <p><sbdl-button class ="primary" *ngIf="isAuthenticated" (click)="logout()">Lgout</sbdl-button></p>
-        <p *ngIf="isAuthenticated">{{userInitials}}</p>
-    </div>
-    <div class="router-outlet">
-        <router-outlet></router-outlet>
-    </div>
+  <sbdl-navigation></sbdl-navigation>
+  <div class="router-outlet"><router-outlet></router-outlet></div>
+  <sbdl-footer></sbdl-footer>
 </div>
