+++ webapp/src/app/layout/app-container/app-container.component.scss
@@ -1,33 +1,32 @@
 @import "spacing";
 
 .app-container {
-    display: grid;
-    grid-template: "navigation router-outlet" 1fr / 64px 1fr;
+  display: grid;
+  grid-template: "navigation router-outlet" "navigation footer" 1fr / 64px 1fr;
 
-    .navigation {
-        grid-area: navigation;
-        background: #5AC923; // todo: define color palette 
-        position: fixed;
-        width: 64px;
-        top: 0%;
-        bottom: 0%;
-        display: flex;
-        flex-direction: column;
-        
-    }
-    .router-outlet {
-        grid-area: router-outlet;
-    }
+  sbdl-navigation {
+    grid-area: navigation;
+  }
+
+  .router-outlet {
+    grid-area: router-outlet;
+  }
+
+  sbdl-footer {
+    grid-area: footer;
+    max-width: 100%;
+  }
 }
 
 @media only screen and (max-width: $breakpoint-sm) {
-    .app-container {
-        grid-template: "navigation" "router-outlet" 1fr / 1fr;
-        .navigation {
-            position: relative;
-            flex-direction: row;
-            justify-content: space-around;
-            width: 100%;
-        }
+  .app-container {
+    grid-template: "navigation" "router-outlet" "footer" 1fr / 1fr;
+
+    sbdl-navigation {
+      position: relative;
+      flex-direction: row;
+      justify-content: space-around;
+      width: 100%;
     }
+  }
 }
