+++ webapp/src/app/layout/app-container/app-container.component.spec.ts
@@ -3,6 +3,8 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { ActivatedRoute } from '@angular/router';
 import { RouterTestingModule } from '@angular/router/testing';
 import { HeaderComponent } from '../header/header.component';
+import { FooterComponent } from '../footer/footer.component';
+import { NavigationComponent } from '../navigation/navigation.component';
 import { AppContainerComponent } from './app-container.component';
 import { mockRootActivatedRouteSnapshot } from 'src/app/app.module.spec';
 import { SbdlCommonModule } from 'src/app/common/common.module';
@@ -22,7 +24,7 @@ describe('AppContainerComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ AppContainerComponent, HeaderComponent ],
+      declarations: [ AppContainerComponent, HeaderComponent, FooterComponent, NavigationComponent ],
       imports: [ RouterTestingModule, SbdlCommonModule ],
       providers: [
         { provide: ActivatedRoute, useValue: mockRootActivatedRouteSnapshot },
