+++ webapp/src/app/layout/layout.module.ts
@@ -3,12 +3,16 @@ import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
 import { AppContainerComponent } from './app-container/app-container.component';
 import { HeaderComponent } from './header/header.component';
+import { FooterComponent } from './footer/footer.component';
+import { NavigationComponent } from './navigation/navigation.component';
 import { SbdlCommonModule } from '../common/common.module';
 
 @NgModule({
   declarations: [
     AppContainerComponent,
-    HeaderComponent
+    HeaderComponent,
+    FooterComponent,
+    NavigationComponent
   ],
   imports: [
     SbdlCommonModule,
