--- webapp/src/app/resource/components/footer/footer.component.html
@@ -1,48 +1,0 @@
-<div class=footer>
-  <div class="footer-links">
-    <div class="member-resources">
-      <h6>California Resources</h6>
-      <ul>
-        <li><a href="#">Reporting System</a></li>
-        <li><a href="#">Practice Test</a></li>
-        <li><a href="#">Assessment Portal</a></li>
-        <li><a href="#">Starting Smarter</a></li>
-        <li><a href="#">Need help?</a></li>
-      </ul>
-    </div>
-    <div class="more-tools-for-teachers">
-      <h6>More Tools for Teachers</h6>
-      <ul>
-        <li><a href="https://csetool.thesnug.dev/">Content Explorer</a></li>
-        <li><a href="https://sampleitems.smarterbalanced.org">Sample Items Website</a></li>
-        <li><a href="https://www.smarterbalanced.org/educators/smarter-balanced-ready/">Big Ideas for Implementation</a></li>
-        <li><a href="https://www.smarterbalanced.org/educators/classroom/">Smarter Balanced in the Classroom</a></li>
-        <li><a href="https://www.smarterbalanced.org/about">Get Involved!</a></li>
-      </ul>
-    </div>
-    <div class="connect">
-      <h6>Connect</h6>
-      <ul>
-        <li><a href="https://twitter.com/SmarterBalanced"><i class="far fa-twitter"></i>Twitter</a></li>
-        <li><a href="https://www.youtube.com/user/smarterbalanced"><i class="far fa-youtube"></i>Facebook</a></li>
-        <li><a href="https://www.youtube.com/user/smarterbalanced"><i class="far fa-youtube"></i>Instagram</a></li>
-        <li><a href="https://www.youtube.com/user/smarterbalanced"><i class="far fa-youtube"></i>YouTube</a></li>
-        <li><a href="https://www.youtube.com/user/smarterbalanced"><i class="far fa-youtube"></i>Pinterest</a></li>
-      </ul>
-    </div>
-    <div class="policy">
-      <h6>Policy</h6>
-      <ul>
-        <li><a href="#">Privacy</a></li>
-        <li><a href="#">Terms & Conditions</a></li>
-        <li><a href="#">Accessibility</a></li>
-      </ul>
-    </div>
-  </div>
-  <hr/>
-  <div class=copyright>
-    <img srg="/assets/svg/smarter-balanced-footer-logo.svg">
-    <img srg="/assets/images/caaspp-footer-logo.png">
-    <span class="copyright">© 2019 The Regents of the University of California</span>
-  </div>
-</div>
