+++ webapp/src/app/resource/instructional/content/instructional-content.component.html
@@ -20,5 +20,3 @@
 <hr *ngIf="resource.thingsToConsider"/>
 
 <sbdl-formative [resource]="resource" (sectionElementLoaded)="setFormative($event)"></sbdl-formative>
-
-<!--<sbdl-footer></sbdl-footer>-->
