+++ webapp/src/app/resource/professional/content/professional-content.component.html
@@ -33,5 +33,3 @@
 <sbdl-formative-list [strategies]="resource.formativeAssessmentStrategies"
                      showProfLearningContent="true"
                      (sectionElementLoaded)="setFormative($event)"></sbdl-formative-list>
-
-<!-- <sbdl-footer></sbdl-footer>-->
