+++ webapp/src/app/resource/resource.component.scss
@@ -21,8 +21,8 @@
 
         background-image: url('/assets/images/blue-spot-2.png');
         background-repeat: no-repeat;
-        background-attachment: fixed;
-        background-position: bottom left 64px;
+        background-attachment: scroll;
+        background-position: bottom left;
     }
 
     .outline:hover {
@@ -58,7 +58,7 @@
 
         background-image: url('/assets/images/green-spot-2.png'), url('/assets/images/yellow-spot-2.png');
         background-repeat: no-repeat;
-        background-attachment: fixed;
+        background-attachment: scroll;
         background-position: right 184px bottom, bottom right;
         min-height: 100vh;
         background-color: $white;
@@ -66,11 +66,9 @@
     }
 
     nav sbdl-outline,
-    aside sbdl-detailed-metadata, // TODO
     aside sbdl-resource-properties {
       position: fixed;
       max-height: 100%;
-      height: 100%;
       overflow-y: auto;
       overflow-x: hidden; // reading mode intentionally clips.
       transition-property: width, min-width;
