+++ webapp/src/sprites/_sprite.scss
@@ -248,3 +248,13 @@
 	height: 24px;
 }
 
+.svg-tft-logo-full-dims {
+	width: 100px;
+	height: 44px;
+}
+
+.svg-tft-logo-small-dims {
+	width: 24px;
+	height: 26px;
+}
+
